requirements.txt
pdfplumber
PyPDF2
pandas
openpyxl
requests
beautifulsoup4
PyYAML
pytest